Clean up walproposer_sim code
