+++ safekeeper/tests/walproposer_sim/mod.rs
@@ -2,6 +2,7 @@ pub mod block_storage;
 pub mod log;
 pub mod safekeeper;
 pub mod safekeeper_disk;
-pub mod util;
+pub mod simulation;
+pub mod simulation_logs;
 pub mod walproposer_api;
 pub mod walproposer_disk;
